# runner/docs/swagger.yaml
basePath: /
definitions:
  api.AsyncDownloadAttachmentsRequest:
    properties:
      debug_mode:
        type: boolean
      gov_ids:
        items:
          type: string
        type: array
    type: object
  api.AsyncDownloadAttachmentsResponse:
    properties:
      error:
        type: string
      errors:
        items:
          type: string
        type: array
      gov_ids:
        items:
          type: string
        type: array
      message:
        type: string
      queued:
        type: integer
      request_ids:
        items:
          type: string
        type: array
      success:
        type: boolean
    type: object
  api.AsyncPipelineRequest:
    properties:
      debug_mode:
        type: boolean
      gov_ids:
        items:
          type: string
        type: array
      intermediate_source:
        $ref: '#/definitions/pipelines.IntermediateSource'
    type: object
  api.AsyncPipelineResponse:
    properties:
      error:
        type: string
      errors:
        items:
          type: string
        type: array
      gov_ids:
        items:
          type: string
        type: array
      message:
        type: string
      queued:
        type: integer
      request_ids:
        items:
          type: string
        type: array
      success:
        type: boolean
    type: object
  api.BulkQueueRequest:
    properties:
      debug_mode:
        type: boolean
      intermediate_source:
        $ref: '#/definitions/pipelines.IntermediateSource'
      limit:
        type: integer
    type: object
  api.BulkQueueResponse:
    properties:
      error:
        type: string
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      queued:
        type: integer
      queued_gov_ids:
        items:
          type: string
        type: array
      success:
        type: boolean
      total_missing:
        type: integer
    type: object
  api.HealthResponse:
    properties:
      services:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
      timestamp:
        type: string
    type: object
  api.QueueClearResponse:
    properties:
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  api.QueueStatusResponse:
    properties:
      queue_stats:
        additionalProperties: true
        type: object
    type: object
  pipelines.IntermediateSource:
    enum:
    - none
    - html
    - raw_json
    - processed_json
    type: string
    x-enum-comments:
      IntermediateSourceHTML: Retrieve from HTML snapshots in S3/local storage
      IntermediateSourceNone: No intermediate source, run full scraping
      IntermediateSourceProcessedJSON: Retrieve from processed JSON objects storage
      IntermediateSourceRawJSON: Retrieve from raw JSON objects storage
    x-enum-descriptions:
    - No intermediate source, run full scraping
    - Retrieve from HTML snapshots in S3/local storage
    - Retrieve from raw JSON objects storage
    - Retrieve from processed JSON objects storage
    x-enum-varnames:
    - IntermediateSourceNone
    - IntermediateSourceHTML
    - IntermediateSourceRawJSON
    - IntermediateSourceProcessedJSON
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing data scraping, processing, and upload pipelines for
    government documents
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Dokito Job Processing API
  version: "1.0"
paths:
  /api/attachments/async:
    post:
      consumes:
      - application/json
      description: Queue download attachments tasks for multiple government IDs for
        background processing
      parameters:
      - description: Async download attachments request with list of government IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AsyncDownloadAttachmentsRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.AsyncDownloadAttachmentsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.AsyncDownloadAttachmentsResponse'
      summary: Execute async download attachments for multiple IDs
      tags:
      - attachments
  /api/health:
    get:
      consumes:
      - application/json
      description: Get the health status of the API and all configured services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.HealthResponse'
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check endpoint
      tags:
      - health
  /api/pipeline/async:
    post:
      consumes:
      - application/json
      description: Queue pipeline tasks for multiple government IDs for background
        processing
      parameters:
      - description: Async pipeline request with list of government IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.AsyncPipelineRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.AsyncPipelineResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.AsyncPipelineResponse'
      summary: Execute async pipeline for multiple IDs
      tags:
      - pipeline
  /api/pipeline/bulk-queue:
    post:
      consumes:
      - application/json
      description: Find missing govids, randomize them, and queue a limited number
        for background processing
      parameters:
      - description: Bulk queue request with optional limit and intermediate source
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.BulkQueueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.BulkQueueResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.BulkQueueResponse'
      summary: Bulk queue missing govids
      tags:
      - pipeline
  /api/queue/clear:
    delete:
      description: Remove all pending, scheduled, and retry tasks from the queue
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.QueueClearResponse'
        "405":
          description: Method Not Allowed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.QueueClearResponse'
      summary: Clear queue
      tags:
      - queue
  /api/queue/status:
    get:
      description: Get the current status of the background task queues
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.QueueStatusResponse'
      summary: Get queue status
      tags:
      - queue
schemes:
- http
- https
swagger: "2.0"